Batch script for sv2v; CV32E40X ABANDONED DUE TO POOR QUARTUS SV SUPPORT
Quartus Prime (Standard and Lite) does not support SystemVerilog-2005 to the degree necessary for compilation of cv32e40x. Converting the rtl code to Verilog-2005 using sv2v is not feasible without significant debug effort and introduces a likely point of failure in our synthesis pipeline. Of the options available, abandoning attempts to use SystemVerilog cores that do not explicitly support Quartus Prime is the least-worst option.